runtime-sdk/evm/precompile: Split standard precompiles into submodules
